requirements.txt
pydantic>=2.0.0
openai>=1.0.0
transformers>=4.30.0
torch
pandas
streamlit
plotly
rouge-score
nltk
tqdm
python-dotenv
pyyaml
fastapi
uvicorn
